final version

+++ src/main/java/com/operation/AdministratorOp.java
@@ -20,6 +20,7 @@ public class AdministratorOp {
     ManagerMapper managerMapper = sqlSession.getMapper(ManagerMapper.class);
     TakesMapper takesMapper = sqlSession.getMapper(TakesMapper.class);
     NecessityMapper necessityMapper = sqlSession.getMapper(NecessityMapper.class);
+    DepartmentMapper departmentMapper = sqlSession.getMapper(DepartmentMapper.class);
 
     public Employee getAdminById(Long employeeId){
         if(administratorMapper.getAdministratorById(employeeId) == null){
@@ -30,17 +31,25 @@ public class AdministratorOp {
     }
 
     // ad (1 增加
-    public void addEmployee(Employee employee){
+    public boolean addEmployee(Employee employee){
+        String department = departmentMapper.getNameById(employee.getDept());
+        if(department == null){
+            return false;
+        }
         administratorMapper.addEmployee(employee);
         sqlSession.commit();
+        return true;
     }
 
     // ad (1 删除
-    public void deleteEmployee(Long employeeId){
+    public boolean deleteEmployee(Long employeeId){
         // 删除教员
         // 1、设置对应课程的教员号为全0
         // 2、删除instructor数据
         // 3、删除employee数据
+        if(employeeMapper.getEmployeeById(employeeId) == null){
+            return false;
+        }
         Instructor instructor = instructorMapper.getInstructorById(employeeId);
         if(instructor != null){
             courseMapper.updateInstructorOfCourse(employeeId);
@@ -60,59 +69,47 @@ public class AdministratorOp {
         employeeMapper.deleteEmployee(employeeId);
         takesMapper.deleteTakesRecord(employeeId);
         sqlSession.commit();
+        return true;
     }
 
-    // ad (1 修改
-    public void updateEmployee(Long employeeId, String section, String update){
-        Employee employee = employeeMapper.getEmployeeById(employeeId);
-        // 系统管理员只允许修改姓名、地址、号码与邮箱
-        switch (section) {
-            case "name" -> employee.setName(update);
-            case "address" -> employee.setAddress(update);
-            case "telephone" -> employee.setTelephone(Long.parseLong(update));
-            case "email" -> employee.setEmail(update);
-        }
-        administratorMapper.updateEmployee(employee);
-        sqlSession.commit();
-    }
-
-    // ad (1 查找
-    public List<Employee> queryEmployee(String searchString){
-        return administratorMapper.searchEmployee(searchString);
+    // ad (2 查找：显示全部
+    public List<Employee> queryAllEmployee(){
+        return administratorMapper.queryAllEmployee();
     }
 
     // ad (2 增加
-    public void addCourse(Course course){
+    public boolean addCourse(Course course){
+        Long instructorId = course.getInstructor_id();
+        if(instructorMapper.getInstructorById(instructorId) == null){
+            return false;
+        }
         courseMapper.addCourse(course);
         sqlSession.commit();
+        return true;
     }
 
     // ad (2 删除
-    public void deleteCourse(Long courseId){
+    public boolean deleteCourse(Long courseId){
         // 1、删除necessity数据
         // 2、删除course数据
+        if(courseMapper.getCourseById(courseId) == null){
+            return false;
+        }
         courseMapper.deleteCourse(courseId);
         necessityMapper.deleteNecessity(courseId);
         sqlSession.commit();
+        return true;
     }
 
     // ad (2 修改
-    public void updateCourse(Long courseId, String section, String update){
-        Course course = courseMapper.getCourseById(courseId);
-        // 系统管理员只能修改课程名、员工号、课程类型与课程内容
-        switch (section){
-            case "course_name" -> course.setCourse_name(update);
-            case "instructor_id" -> course.setInstructor_id(Integer.parseInt(update));
-            case "type" -> course.setType(update);
-            case "content" -> course.setContent(update);
-        }
+    public void updateCourse(Course course){
         administratorMapper.updateCourse(course);
         sqlSession.commit();
     }
 
     // ad (2 查找
-    public List<Course> queryCourse(String searchString){
-        return administratorMapper.searchCourse(searchString);
+    public List<Course> queryAllCourse(){
+        return administratorMapper.queryAllCourse();
     }
 
     // ad (3 根据员工号查询用户的个人基本信息与培训成绩信息
@@ -136,18 +133,22 @@ public class AdministratorOp {
     }
 
     // ad (4 管理日志信息：默认只能查看日志与增加日志，不能删除或修改日志
-    public List<Log> queryLog(String searchString){
-        return administratorMapper.queryLog(searchString);
-    }
-
-    public void addLog(Long a_id, String operation){
+    public boolean addLog(Long employee_id, String operation){
+        if(employeeMapper.getEmployeeById(employee_id) == null){
+            return false;
+        }
         Log log = new Log();
-        log.setA_id(a_id);
+        log.setEmployee_id(employee_id);
         log.setOperation(operation);
         Date date = new Date();
         log.setDate(new Timestamp(date.getTime()));
         administratorMapper.addLog(log);
         sqlSession.commit();
+        return true;
+    }
+
+    public List<Log> queryAllLog(){
+        return administratorMapper.queryAllLog();
     }
 
     public void close(){
